requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.26.2
gdown==4.7.1
python-dotenv==1.0.0 